refactor: 파일명 및 경로변경 items... -> products... 로 통일
