+++ react/my-react-app/src/App.jsx
@@ -4,7 +4,7 @@ import Layout from "./components/Layout";
 import HomePage from "./pages/HomePage";
 import SigninPage from "./pages/SigninPage";
 import SignupPage from "./pages/SignupPage";
-import ItemsPage from "./pages/ItemsPage";
+import ProductsPage from "./pages/ProductsPage";
 import AddItemPage from "./pages/AddItemPage";
 import PrivacyPage from "./pages/PrivacyPage";
 import FaqPage from "./pages/FaqPage";
@@ -22,8 +22,8 @@ function App() {
         {/* Main routes with layout */}
         <Route element={<Layout />}>
           <Route path="/" element={<HomePage />} />
-          <Route path="/items" element={<ItemsPage />} />
-          <Route path="/items/:productId" element={<ProductDetailPage />} />
+          <Route path="/products" element={<ProductsPage />} />
+          <Route path="/products/:productId" element={<ProductDetailPage />} />
           <Route path="/additem" element={<AddItemPage />} />
           <Route path="/privacy" element={<PrivacyPage />} />
           <Route path="/faq" element={<FaqPage />} />
